Smarter (#35)
* Rename to group related files.

* Implement as_* functions.

* Remove api_ prefixes.

* Remove noisy TODOs.

* Return *all* the length issues, not just the first one.

Closes #33.

* Implement as_license.

* Don't check urls.

The regex doesn't actually work, and these urls can be non-standard. We don't really care about these until we're using them, let's let people do weird things here.

* Scrub out apid_list.

And distribute TODOs.

* Implement as_info()

* Standardize server_variable_list() to variables()

Closes #30.

This is the only place that variables() appears in the definition, and there's really no reason that it would have to be server-specific anyway.

* Implement as_*().

* Rename server_variable to string_replacements.

I couldn't stick with the OpenAPI names. This makes much more sense to me. More to come in this area.

* Rename "variables" to "server_variables".

This is a list of "string_replacements" objects, which are themselves lists. I want to make sure these stand out as special for a multi-server-with-different-variables situation.

* Finish as_server() implementation.

* Clean docs.

* Implement as_rapid().

* Delete unused test.

* Styler

* Add missing dot help.